requirements.txt
fastapi
uvicorn[standard]
elasticsearch
psycopg2-binary
pypika
pydantic-settings
